requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.11.0
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.7
cryptography==40.0.2
dnspython==2.3.0
docstring-parser==0.15
ecdsa==0.18.0
email-validator==2.0.0.post2
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.95.2
firebase-admin==6.2.0
frozenlist==1.4.0
google==3.0.0
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.56.0
grpcio-status==1.56.0
h11==0.14.0
httpcore==0.17.0
httplib2==0.22.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
Mako==1.2.4
MarkupSafe==2.1.2
minio==7.1.15
msgpack==1.0.5
multidict==6.0.4
mysqlclient==2.1.1
numpy==1.25.0
openpyxl==3.1.2
orjson==3.8.12
packaging==23.1
pandas==2.0.2
passlib==1.7.4
pluggy==1.0.0
prometheus-client==0.17.0
proto-plus==1.22.3
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
pyhumps==3.8.0
PyJWT==2.7.0
pyparsing==3.1.0
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
requests==2.30.0
rsa==4.9
schedule==1.2.0
six==1.16.0
slugify==0.0.1
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.13
starlette==0.27.0
text-unidecode==1.3
tomli==2.0.1
twilio==8.5.0
typing_extensions==4.5.0
tzdata==2023.3
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
